fix build cache

+++ scripts/libbluray/build.sh
@@ -1,9 +1,13 @@
-#!/bin/sh
+#!/bin/bash
 
 set -e # exit immediately if a command exits with a non-zero status
 set -u # treat unset variables as an error
 
 cd ${SRC_DIR}
+if [[ -d ${OUTPUT_DIR} ]];then
+    echo "already exists,skip"
+    exit 0
+fi
 rm -rf contrib/libudfread
 tar -zxf ${DOWNLOADS_DIR}/libudfread-1.2.0.tar.gz -C contrib
 mv contrib/libudfread-1.2.0 contrib/libudfread
