+++ scripts/libplacebo/build.sh
@@ -1,9 +1,13 @@
-#!/bin/sh
+#!/bin/bash
 
 set -ex # exit immediately if a command exits with a non-zero status
 set -u # treat unset variables as an error
 
 cd ${SRC_DIR}
+if [[ -d ${OUTPUT_DIR} ]];then
+    echo "already exists,skip"
+    exit 0
+fi
 
 rm -rf ./3rdparty
 mkdir -p ./3rdparty 
@@ -24,13 +28,18 @@ mv ./3rdparty/MarkupSafe-2.1.2 ./3rdparty/markupsafe
 tar -xzvf "${DOWNLOADS_DIR}/Vulkan-Headers-1.4.336.tar.gz" -C ./3rdparty/
 mv ./3rdparty/Vulkan-Headers-1.4.336 ./3rdparty/Vulkan-Headers
 
+cp ${PROJECT_DIR}/cross-files/${OS}-${ARCH}.ini /tmp/${OS}-${ARCH}.ini
+if grep '\-std=c++14' /tmp/${OS}-${ARCH}.ini;then
+    sed -i 's/-std=c++14/-std=c++17/g'  /tmp/${OS}-${ARCH}.ini
+fi 
+
 meson setup build \
-    --cross-file ${PROJECT_DIR}/cross-files/${OS}-${ARCH}.ini \
+    --cross-file /tmp/${OS}-${ARCH}.ini \
     --prefix="${OUTPUT_DIR}" \
     -Ddemos=false \
     -Dtests=false \
     -Dvulkan=enabled \
-    --default-library=static 
+    --default-library=static
 
 meson compile -C build
 meson install -C build
